removing player names from hash creation
The space dedicated to p1 and p2 names on the replay file is not wiped between matches, only overwritten, so if you play against someone with a longer name than your current opponent your p1 and p2 names will not be the same in the replay creating different hashes for the "same" replay.  I could look for null terminations but it's not worth it.